Fixed the type check
